Fix tests and extract common test util code
